test(reader): split module picker coverage
Why:
- Continue the test architecture migration by moving BibleUI-owned reader module picker coverage out of the app-host unit bundle.
- Keep Android ChooseDocument parity tests in the fast package test lane instead of paying app bootstrap and simulator install cost.

What Changed:
- Added BibleReaderModulePickerTests under the BibleUI package test target with the migrated filtering, pseudo-document, document-management, routing, and full-screen chooser guardrails.
- Removed the matching module-picker tests from AndBibleTests+AppAndReader while documenting the remaining app-host scene-configuration sentinel.
- Updated the migration plan with the module-picker slice and current AppAndReader remainder.

Validation:
- python3 scripts/check_repo_standards.py docblocks --all-files
- python3 scripts/check_repo_standards.py source-guards
- git diff --cached --check
- xcodebuild test -project AndBible.xcodeproj -scheme BibleUITests -destination platform=iOS Simulator,name=iPhone 17 -derivedDataPath /tmp/andbible-module-picker-bibleui-dd2 CODE_SIGNING_ALLOWED=NO -only-testing:BibleUITests/BibleReaderModulePickerTests
- xcodebuild test -quiet -project AndBible.xcodeproj -scheme AndBible -destination platform=iOS Simulator,name=iPhone 17 -derivedDataPath /tmp/andbible-module-picker-app-dd2 CODE_SIGNING_ALLOWED=NO -only-testing:AndBibleTests/AndBibleTests/testApplicationDelegateSceneConfigurationUsesWindowSceneDelegate
- GitNexus staged detect_changes reported no mapped production-flow changes.
- Adversarial review findings were addressed before commit.

Impact:
- Reduces app-host test surface without changing production behavior.
- Preserves Android document chooser parity coverage in the owning package target.
