+++ docs/test-architecture-migration-plan.md
@@ -158,7 +158,8 @@ Blocker: `AndBibleTestSupport.swift` is a 2,233-line `extension AndBibleTests` e
 - `AndBibleTests+AppAndReader` module-switch planner slice has moved to `BibleUITests` because it exercises BibleUI current-document switch planning, Android category mismatch rejection, atomic auxiliary-document persistence, module-only selection, and category reload decisions without SWORD fixtures or app bootstrap.
 - `AndBibleTests+AppAndReader` reader chrome slice has moved to `BibleUITests` because it exercises BibleUI header, toolbar action, popup placement, overflow menu, Downloads routing, and keyboard shortcut construction without app delegate/bootstrap behavior.
 - `AndBibleTests+AppAndReader` module-browser slice has been split by owning module: Downloads filtering/sorting, failed-source cache merge, and startup-default installable-row guards moved to `BibleUITests`; Android metadata token decoding and catalog `InstallSize` byte preservation moved to `SwordKitTests`; the duplicate recommended-document refresh failure test was removed from the app-host bundle because equivalent SwordKit package coverage already exists in `SwordManagerTests`.
-- Batches for the genuinely BibleUI-behavior remainder: ReaderNavigation, Bookmarks/Strongs, Window/Settings, view-model parts of `+AppAndReader`.
+- `AndBibleTests+AppAndReader` module-picker slice has moved to `BibleUITests` because it exercises BibleUI `ChooseDocument` filtering, pseudo-document/add-on rows, document-management actions, category mapping, map routing, and Android full-screen chooser presentation without app delegate/bootstrap behavior.
+- Current `+AppAndReader` remainder: quick-module selector/document-switch source guards, reader pane active-window styling, speak/drawer shell construction, passage chooser source/progress contracts, color helpers, and the app-hosted scene-configuration sentinel. Move each package-owned test to the lowest owning package target; keep only true app-bootstrap coverage in `AndBibleTests`.
 - **Split `+AppAndReader` deliberately**: `sceneConfiguration` test stays app-hosted; the `ContentView` legacy root-sidebar source scan is now a repo-standards `source-guards` check; the remaining package-owned tests should move to the lowest owning package target.
 - Run BibleUITests via `xcodebuild test -scheme BibleUITests -destination 'platform=iOS Simulator,...'` against the **committed package test scheme from Phase 0** (no app host).
 - Gate per batch: moved batch green in new target; equal count removed from app bundle; app build still green.
